Tweak user live files
