+++ lib/aliancer_web/live/user_live/forgot_password.ex
@@ -1,4 +1,4 @@
-defmodule AliancerWeb.UserConfirmationInstructionsLive do
+defmodule AliancerWeb.UserLive.ForgotPassword do
   use AliancerWeb, :live_view
 
   alias Aliancer.Accounts
@@ -7,20 +7,19 @@ defmodule AliancerWeb.UserConfirmationInstructionsLive do
     ~H"""
     <div class="mx-auto max-w-sm">
       <.header class="text-center">
-        No confirmation instructions received?
-        <:subtitle>We'll send a new confirmation link to your inbox</:subtitle>
+        Forgot your password?
+        <:subtitle>We'll send a password reset link to your inbox</:subtitle>
       </.header>
 
-      <.simple_form for={@form} id="resend_confirmation_form" phx-submit="send_instructions">
+      <.simple_form for={@form} id="reset_password_form" phx-submit="send_email">
         <.input field={@form[:email]} type="email" placeholder="Email" required />
         <:actions>
           <.button phx-disable-with="Sending..." class="w-full">
-            Resend confirmation instructions
+            Send password reset instructions
           </.button>
         </:actions>
       </.simple_form>
-
-      <p class="text-center mt-4">
+      <p class="text-center text-sm mt-4">
         <.link href={~p"/users/register"}>Register</.link>
         | <.link href={~p"/users/log_in"}>Log in</.link>
       </p>
@@ -32,16 +31,16 @@ defmodule AliancerWeb.UserConfirmationInstructionsLive do
     {:ok, assign(socket, form: to_form(%{}, as: "user"))}
   end
 
-  def handle_event("send_instructions", %{"user" => %{"email" => email}}, socket) do
+  def handle_event("send_email", %{"user" => %{"email" => email}}, socket) do
     if user = Accounts.get_user_by_email(email) do
-      Accounts.deliver_user_confirmation_instructions(
+      Accounts.deliver_user_reset_password_instructions(
         user,
-        &url(~p"/users/confirm/#{&1}")
+        &url(~p"/users/reset_password/#{&1}")
       )
     end
 
     info =
-      "If your email is in our system and it has not been confirmed yet, you will receive an email with instructions shortly."
+      "If your email is in our system, you will receive instructions to reset your password shortly."
 
     {:noreply,
      socket
