+++ lib/aliancer_web/router.ex
@@ -92,10 +92,10 @@ defmodule AliancerWeb.Router do
 
     live_session :redirect_if_user_is_authenticated,
       on_mount: [{AliancerWeb.UserAuth, :redirect_if_user_is_authenticated}] do
-      live "/users/register", UserRegistrationLive, :new
-      live "/users/log_in", UserLoginLive, :new
-      live "/users/reset_password", UserForgotPasswordLive, :new
-      live "/users/reset_password/:token", UserResetPasswordLive, :edit
+      live "/users/register", UserLive.Registration, :new
+      live "/users/log_in", UserLive.Login, :new
+      live "/users/reset_password", UserLive.ForgotPassword, :new
+      live "/users/reset_password/:token", UserLive.ResetPassword, :edit
     end
 
     post "/users/log_in", UserSessionController, :create
@@ -106,8 +106,8 @@ defmodule AliancerWeb.Router do
 
     live_session :require_authenticated_user,
       on_mount: [{AliancerWeb.UserAuth, :ensure_authenticated}] do
-      live "/users/settings", UserSettingsLive, :edit
-      live "/users/settings/confirm_email/:token", UserSettingsLive, :confirm_email
+      live "/users/settings", UserLive.Settings, :edit
+      live "/users/settings/confirm_email/:token", UserLive.Settings, :confirm_email
     end
   end
 
@@ -118,8 +118,8 @@ defmodule AliancerWeb.Router do
 
     live_session :current_user,
       on_mount: [{AliancerWeb.UserAuth, :mount_current_user}] do
-      live "/users/confirm/:token", UserConfirmationLive, :edit
-      live "/users/confirm", UserConfirmationInstructionsLive, :new
+      live "/users/confirm/:token", UserLive.Confirmation, :edit
+      live "/users/confirm", UserLive.ConfirmationInstructions, :new
     end
   end
 end
